+++ dev/lexer/lexer.c
@@ -6,28 +6,17 @@
 /*   By: akurochk <akurochk@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/16 18:13:05 by akurochk          #+#    #+#             */
-/*   Updated: 2024/04/25 14:06:49 by akurochk         ###   ########.fr       */
+/*   Updated: 2024/04/25 14:43:17 by akurochk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../inc/minishell.h"
 
-static int	is_space(const char c)
-{
-	return ((c >= 9 && c <= 13) || c == ' ');
-}
-
-static int	is_operator(const char c, const char c2)
-{
-	return (c == '(' || c == ')' || c == '|'
-		|| (c == '&' && c2 == '&') || c == '<' || c == '>');
-}
-
 /*
 Returns the position of the second quotes relative to the first quotes.
 Returns 0 if didn't find second quotes.
 */
-long	lexer_where_second_quotes(char c, const char *line)
+static long	lexer_where_second_quotes(char c, const char *line)
 {
 	long	i;
 
@@ -41,7 +30,7 @@ long	lexer_where_second_quotes(char c, const char *line)
 /*
 Returns 0 if OK.
 */
-int	lexer_get_quotes(t_list *toks, const char *line, long i, long *end)
+static int	lexer_get_quotes(t_list *toks, const char *line, long i, long *end)
 {
 	char	*s_str;
 
@@ -71,35 +60,9 @@ int	lexer_get_quotes(t_list *toks, const char *line, long i, long *end)
 }
 
 /*
-Returns token type.
-*/
-long	get_operator_type(const char *line, long i)
-{
-	if (line[i] == '(')
-		return (L_PAR_L);
-	if (line[i] == ')')
-		return (L_PAR_R);
-	if (line[i] == '|' && line[i + 1] == '|')
-		return (L_OR);
-	if (line[i] == '|')
-		return (L_PIPE);
-	if (line[i] == '&' && line[i + 1] == '&')
-		return (L_AND);
-	if (line[i] == '<' && line[i + 1] == '<')
-		return (L_RE_DOC);
-	if (line[i] == '>' && line[i + 1] == '>')
-		return (L_RE_APP);
-	if (line[i] == '<')
-		return (L_RE_IN);
-	if (line[i] == '>')
-		return (L_RE_OUT);
-	return (L_ANOTH);
-}
-
-/*
 Returns 0 if OK.
 */
-int	lexer_get_operator(t_list *toks, const char *line, long i, long *end)
+static int	lexer_get_op(t_list *toks, const char *line, long i, long *end)
 {
 	long	t;
 
@@ -107,34 +70,11 @@ int	lexer_get_operator(t_list *toks, const char *line, long i, long *end)
 	if (t == L_ANOTH)
 		return (errors(1, "Error: get_operator_type", 0, 258));
 	if (!list_put(toks, (int *)t, NULL))
-		return (errors(1, "Erorr: lexer_get_operator", 1, 0));
+		return (errors(1, "Erorr: lexer_get_op", 1, 0));
 	if (t == L_OR || t == L_AND || t == L_RE_DOC || t == L_RE_APP)
 		*end = 1;
 	else
 		*end = 0;
-	return (0);
-}
-
-/*
-Returns 0 if OK.
-*/
-int	lexer_get_word(t_list *toks, const char *line, long i, long *end)
-{
-	char	*s_str;
-
-	*end = 0;
-	while (!is_space(line[i + *end])
-		&& !is_operator(line[i + *end], line[i + *end + 1])
-		&& line[i + *end] != '\'' && line[i + *end] != '\"' && line[i + *end])
-		*end = *end + 1;
-	s_str = ft_substr(line, i, *end);
-	if (!s_str)
-		return (1);
-	if (!list_put(toks, (int *)L_WORD, s_str))
-	{
-		free (s_str);
-		return (1);
-	}
 	return (0);
 }
 
@@ -160,7 +100,7 @@ static int	lexer_get_token(t_list *toks, const char *line, long *i)
 	}
 	else if (is_operator(line[*i], line[*i + 1]))
 	{
-		if (lexer_get_operator(toks, line, *i, &end))
+		if (lexer_get_op(toks, line, *i, &end))
 			return (1);
 		*i = *i + end + 1;
 	}
