+++ dev/list/list.c
@@ -6,7 +6,7 @@
 /*   By: akurochk <akurochk@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/16 10:19:32 by akurochk          #+#    #+#             */
-/*   Updated: 2024/04/24 17:04:47 by akurochk         ###   ########.fr       */
+/*   Updated: 2024/04/25 14:17:48 by akurochk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -103,164 +103,4 @@ int	list_replace(t_list *list, void *key, void *new_val)
 	}
 	e_ptr->next = elem_new(key, new_val);
 	return (e_ptr->next != NULL);
-}
-
-/*
-Relinks elements.
-Destructs and free node.
-*/
-static void	elem_del(t_list *list, t_elem *curr, t_elem *prev)
-{
-	if (prev)
-		prev->next = curr->next;
-	else
-		list->head = curr->next;
-	if (list->dst_key)
-		list->dst_key(curr->key);
-	if (list->dst_val)
-		list->dst_val(curr->val);
-	free(curr);
-}
-
-/*
-Returns 0 if list == NULL.
-Returns 1 if node was deleted or if didn't find node with the same key.
-*/
-int	list_del(t_list *list, void *key)
-{
-	t_elem	*e_curr;
-	t_elem	*e_prev;
-
-	if (!list)
-		return (0);
-	if (list_size(list) == 0)
-		return (1);
-	e_curr = list->head;
-	e_prev = NULL;
-	while (e_curr)
-	{
-		if ((list->cmp_key(key, e_curr->key)) == 0)
-		{
-			elem_del(list, e_curr, e_prev);
-			return (1);
-		}
-		e_prev = e_curr;
-		e_curr = e_curr->next;
-	}
-	return (1);
-}
-
-/*
-Returnes list size or 0 for empty list
-*/
-int	list_size(t_list *list)
-{
-	int		size;
-	t_elem	*curr;
-
-	size = 0;
-	if (!list)
-		return (size);
-	curr = list->head;
-	while (curr)
-	{
-		curr = curr->next;
-		size++;
-	}
-	return (size);
-}
-
-/*
-Returns 1 if the list contains the key.
-Returns 0 if not.
-*/
-int	list_contains_key(t_list *list, void *key)
-{
-	t_elem	*e_ptr;
-
-	if (!list || list_size(list) == 0)
-		return (0);
-	e_ptr = list->head;
-	while (e_ptr)
-	{
-		if (list->cmp_key(e_ptr->key, key) == 0)
-			return (1);
-		e_ptr = e_ptr->next;
-	}
-	return (0);
-}
-
-/*
-Returns ptr to a value from the list according the key.
-Returns NULL if list emty or no key in list.
-*/
-void	*list_get(t_list *list, void *key)
-{
-	t_elem	*e_ptr;
-
-	if (!list || list_size(list) == 0)
-		return (NULL);
-	e_ptr = list->head;
-	while (e_ptr)
-	{
-		if (list->cmp_key(e_ptr->key, key) == 0)
-			return (e_ptr->val);
-		e_ptr = e_ptr->next;
-	}
-	return (NULL);
-}
-
-/*
-Destructs keys and values, then frees element, then frees list. 
-*/
-void	list_free(void *list)
-{
-	t_elem	*e_curr;
-	t_elem	*e_next;
-	t_list	*l;
-
-	l = (t_list *)list;
-	if (!list)
-		return ;
-	e_curr = l->head;
-	while (e_curr)
-	{
-		e_next = e_curr->next;
-		if (l->dst_key)
-			l->dst_key(e_curr->key);
-		if (l->dst_val)
-			l->dst_val(e_curr->val);
-		free(e_curr);
-		e_curr = e_next;
-	}
-	free(l);
-}
-
-/*
-Comporators.
-Do we need another comparator?
-*/
-int	cmp_int(const void *key_1, const void *key_2)
-{
-	int	*a;
-	int	*b;
-
-	a = (int *)key_1;
-	b = (int *)key_2;
-	return (*a - *b);
-}
-
-int	ft_strcmp(const char *s1, const char *s2)
-{
-	int	i;
-
-	i = 0;
-	while (s1[i] && s1[i] == s2[i])
-		i++;
-	return ((unsigned char)s1[i] - (unsigned char)s2[i]);
-}
-
-int	cmp_str(const void *key_1, const void *key_2)
-{
-	return (ft_strcmp((const char *)key_1, (const char *)key_2));
 }
