+++ dev/main.c
@@ -6,7 +6,7 @@
 /*   By: akurochk <akurochk@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/09 10:13:18 by akurochk          #+#    #+#             */
-/*   Updated: 2024/04/25 14:06:44 by akurochk         ###   ########.fr       */
+/*   Updated: 2024/04/25 15:29:45 by akurochk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,6 @@
 
 int	g_signal = 0;
 /*
-
-MAIN loop:
-start		-> READLINE	-> line
-line		-> LEXER		-> tokens
-tokens		-> PARSER		-> analaise tokens and get info about builtin(s)
-builtin(s)	-> EXECUTE		-> result
-result		-> POSTACTIONS	-> start/finish
-
 BUILTINS:
 echo	- option -n
 cd		- only a relative or absolute path
@@ -30,73 +22,7 @@ export	- no options
 unset	- no options
 env		- no options or arguments
 exit	- no options
-
 */
-
-int	str_contains(const char *s, const char c)
-{
-	if (s)
-	{
-		while (*s)
-		{
-			if (*s == c)
-				return (1);
-			s++;
-		}
-	}
-	return (0);
-}
-
-char	*ft_strdup(const char *s)
-{
-	char	*ret;
-	int		i;
-
-	if (!s)
-		return (NULL);
-	i = 0;
-	ret = (char *)malloc(sizeof(char) * (ft_strlen(s) + 1));
-	if (!ret)
-		return (NULL);
-	while (s[i])
-	{
-		ret[i] = s[i];
-		i++;
-	}
-	ret[i] = '\0';
-	return (ret);
-}
-
-/*
-Returns 0 if OK.
-Returns 1 if error.
-*/
-int	str_split(const char *s, const char *sep, char **dst1, char **dst2)
-{
-	int	i;
-	int	j;
-
-	if (!s || !sep)
-		return (1);
-	i = 0;
-	while (s[i] && !str_contains(sep, s[i]))
-		i++;
-	j = i;
-	*dst1 = (char *)malloc(sizeof(char) * (i + 1));
-	if (!dst1)
-		return (1);
-	while (j--)
-		(*dst1)[j] = s[j];
-	(*dst1)[i] = '\0';
-	if (!s[i])
-		*dst2 = NULL;
-	else
-		*dst2 = ft_strdup(s + i + 1);
-	if (dst2)
-		return (0);
-	free(*dst1);
-	return (1);
-}
 
 int	env_parser(t_list *env_lst, char **env)
 {
@@ -109,7 +35,7 @@ int	env_parser(t_list *env_lst, char **env)
 	i = -1;
 	while (env[++i])
 	{
-		if (str_split(env[i], "=", &key, &val))
+		if (ft_str_split(env[i], "=", &key, &val))
 			return (errors(1, "Error: env_parser", 1, 0));
 		if (!list_put(env_lst, key, val))
 			return (errors(1, "Error: env_parser", 1, 0));
@@ -134,16 +60,13 @@ int	init_data(t_data *data, char **env)
 	data->builtins[4] = "unset";
 	data->builtins[5] = "env";
 	data->builtins[6] = "exit";
-	// set pointers to builtins functions
-	/*
-	data->f_builtins[0] = &ft_echo;
-	data->f_builtins[1] = &ft_cd;
-	data->f_builtins[2] = &ft_pwd;
-	data->f_builtins[3] = &ft_export;
-	data->f_builtins[4] = &ft_unset;
-	data->f_builtins[5] = &ft_env;
-	data->f_builtins[6] = &ft_exit;
-	*/
+	data->f_builtins[0] = NULL; // &ft_echo;
+	data->f_builtins[1] = NULL; // &ft_cd;
+	data->f_builtins[2] = NULL; // &ft_pwd;
+	data->f_builtins[3] = NULL; // &ft_export;
+	data->f_builtins[4] = NULL; // &ft_unset;
+	data->f_builtins[5] = NULL; // &ft_env;
+	data->f_builtins[6] = NULL; // &ft_exit;
 	return (0);
 }
 
@@ -159,23 +82,16 @@ void	main_exit(t_data *data)
 	rl_clear_history();
 }
 
-char	*ft_readline(void)
-{
-	char	*line;
-
-	line = readline(PROMT);
-	if (line && *line)
-		add_history(line);
-	return (line);
-}
-
 /* new token list, read line, add history */
 /* SIGINT = Ctrl + C */
-void	main_begining(t_list **toks)
+void	main_begining(t_list **toks, char **line)
 {
 	signal(SIGINT, handler_signal);
 	errno = 0;
 	*toks = list_new(cmp_int, NULL, free);
+	*line = readline(PROMT);
+	if (*line && **line)
+		add_history(*line);
 }
 
 void	main_postaction(t_list *toks)
@@ -183,6 +99,28 @@ void	main_postaction(t_list *toks)
 	list_free(toks);
 	signal(SIGQUIT, SIG_IGN);
 }
+
+//============TEST============
+void	test_print_tokens(t_list *toks)
+{
+	if (!toks)
+		printf("\n\033[0;31mtest_print_tokens FAIL\033[0m\n");
+	printf("\n\033[0;34mtest_print_tokens \\\033[0m\n");
+	t_elem *e_ptr = toks->head;
+	long i = 0;
+	while (e_ptr)
+	{
+		long  key = (long)(e_ptr->key);
+		char *val = (char *)(e_ptr->val);
+		printf("Token [%ld]	", i);
+		printf("Key [%ld]		", key);
+		printf("Val [%s]\n", val);
+		e_ptr = e_ptr->next;
+		i++;
+	}
+	printf("\033[0;34mtest_print_tokens /\033[0m\n\n");
+}
+//============TEST============
 
 int	main(int ac, const char **av, char **env)
 {
@@ -196,14 +134,16 @@ int	main(int ac, const char **av, char **env)
 		return (EXIT_FAILURE);
 	while (data.flag_exit)
 	{
-		main_begining(&toks);
-		line = ft_readline();
+		main_begining(&toks, &line);
 		if (line == NULL)
 			data.flag_exit = 0;
 		else if (line != NULL && *line)
 		{
 			if (lexer(line, toks) == 0)
+			{
+				test_print_tokens(toks); //============TEST
 				parser(&data, toks);
+			}
 			free(line);
 		}
 		main_postaction(toks);
